move test to tip20 mod.rs
